Changes:
Made LeaderboardScreen, Sharescreen UI display properly
Made losing the game cause the game to close
Added JavaDoc for all the files in UI (EXCEPT FARM which is not implemented fully yet)

Got rid of unnesscary files inside UI package :
ComponentFactory
Z_Button
Z_GameUI
Z_GameUI_SAVE
Z_Label
Z_Panel
Z_TextField
E_ShopScreen (Which is replaced by E_ShopScreen_New)

fixed a typo inside of LoadGameScreen
