+++ src/main/java/UI/A0_MainScreen.java
@@ -1,24 +1,33 @@
 package UI;
 
-
 import javax.swing.*;
 import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 
+/**
+ *This is the MainScreen that everyone will be using to build their screens
+ *
+ * specifically the other screens will be using the JFrame window, Container con, and titleFont, normalFont
+ *
+ * so that the other screens don't need to initialize the window, container and fonts each time
+ */
 public class A0_MainScreen {
-    //window
+    /**
+     * initializes
+     * 1. JFrame window, which all the screens will be building on top of
+     * 2. Container con, which all the other screens will add their panels to
+     * 3. titleFont for use for the other Screens
+     * 4. normalFont for use for the other Screens
+     */
     JFrame window;
     Container con;
-
-    //fonts
     Font titleFont = new Font("Times New Roman", Font.PLAIN, 96);
     Font normalFont = new Font("Times New Roman", Font.PLAIN, 25);
 
-    //Class Instantiation
+    /**
+     * Settings the parameters for the window
+     */
 
     public A0_MainScreen(){
-        //makes window
         window = new JFrame();
         window.setSize(800,600);
         window.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
@@ -28,6 +37,10 @@ public class A0_MainScreen {
         con = window.getContentPane();
     }
 
+    /**
+     * A quit method that will be called by all the other screens, to use when switching from their screen to another,
+     * to close the screen, so that there is ever only one screen open at the same time
+     */
     public void quit(){
         window.setVisible(false);
         window.dispose();
