+++ src/main/java/UI/A_TitleScreen.java
@@ -5,19 +5,33 @@ import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
+/**
+ * This is the TitleScreen, which you start in when you first start the game
+ *
+ * In this screen the title of the game is displayed, along with a start button, which takes you to the new/load screen
+ */
 public class A_TitleScreen extends A0_MainScreen {
-    //title name
+    /**
+     * initializes title Components
+     */
     JPanel titlePanel;
     JLabel titleLabel;
-    //startButton
+    /**
+     * initializes start button Components
+     */
     JPanel startButtonPanel;
     JButton startButton;
-    //transition from Title Screen to Sign in Screen
-    titleToSignIn tsHandler = new titleToSignIn();
 
+    /**
+     * ActionListener which transitions the Title Screen to new/load screen
+     */
+    titleToNewLoad tsHandler = new titleToNewLoad();
 
-    //A_TitleScreen.TitleScreenHandler tsHandler = new A_TitleScreen.TitleScreenHandler();
-
+    /**
+     * Setting the parameters for the title, and start button
+     *
+     * adding all the panels to the container con, which was initialized in A0_MainScreen
+     */
     public A_TitleScreen(){
         //have to put startButton here for order of layers
         startButton = new JButton("START");
@@ -26,7 +40,7 @@ public class A_TitleScreen extends A0_MainScreen {
         titlePanel = new JPanel();
         titlePanel.setBounds(100,100,600,150);
         titlePanel.setBackground(Color.BLACK);
-        titleLabel = new JLabel("Farming Game", SwingConstants.CENTER);
+        titleLabel = new JLabel("Ranchville", SwingConstants.CENTER);
         titleLabel.setForeground(Color.white);
         titleLabel.setFont(titleFont);
 
@@ -45,25 +59,28 @@ public class A_TitleScreen extends A0_MainScreen {
         con.add(titlePanel);
         con.add(startButtonPanel);
     }
-
-
-
-    public class titleToSignIn implements ActionListener {
+    
+    /**
+     * ActionListener title screen to new/loadscreen
+     */
+    public class titleToNewLoad implements ActionListener {
         public void actionPerformed(ActionEvent event){
             C_NewOrLoadScreen.createAndShowNewOrLoadScreen();
             quit();
         }
     }
 
+    /**
+     * initializes the title screen to be called by main
+     */
     public static void createAndShowTitleScreen() {
-        //Create and set up the content pane.
         A_TitleScreen a = new A_TitleScreen();
-
-        //newContentPane.setOpaque(true);
     }
+
+    /**
+     * when this screen is called, the screen is displayed
+     */
     public static void main(String[] args) {
-        //Schedule a job for the event-dispatching thread:
-        //creating and showing this application's GUI.
         javax.swing.SwingUtilities.invokeLater(new Runnable() {
             public void run() {
                 createAndShowTitleScreen();
