+++ src/main/java/UI/D_NewGameScreen.java
@@ -8,14 +8,34 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.IOException;
 
+/**
+ * if the player chooses to create a new game, they will be able to enter the name of the farm,
+ * which will be saved so that player can choose to load the game later, the player will be sent to the main menu
+ */
 public class D_NewGameScreen extends C_NewOrLoadScreen {
+    /**
+     * initializes text Components, which will prompt the player to enter a name for the farm that they are creating
+     */
     JPanel newGameTextPanel;
-    JPanel newGamePanel;
     JLabel newGameTextLabel;
+
+    /**
+     * initializes text Input Components, which will allow the player to enter the name of their farm
+     */
+    JPanel newGamePanel;
     JTextField jtf;
     JButton enterB;
+
+    /**
+     * ActionListeners which transitions the new game to menu screen
+     */
     newGameToMain mHandler = new newGameToMain();
 
+    /**
+     * Setting the parameters for the text, and text input for the setting the farm name
+     *
+     * adding all the panels to the container con, which was initialized in A0_MainScreen
+     */
     public D_NewGameScreen() {
         newLoadTextPanel.setVisible(false);
         newButtonPanel.setVisible(false);
@@ -48,6 +68,13 @@ public class D_NewGameScreen extends C_NewOrLoadScreen {
         con.add(newGamePanel);
     }
 
+    /**
+     * ActionListener new game screen to menu
+     *
+     *the player will be able to enter the name of the farm,
+     *which will be saved so that player can choose to load the game later
+     *the player will be sent to the main menu
+     */
     public class newGameToMain implements ActionListener {
         public void actionPerformed(ActionEvent event){
             try {
@@ -61,8 +88,10 @@ public class D_NewGameScreen extends C_NewOrLoadScreen {
         }
     }
 
+    /**
+     * initializes the new screen to be called by C_NewOrLoadScreen
+     */
     public static void createAndShowNewGameScreen() {
-        //Create and set up the content pane.
         D_NewGameScreen d = new D_NewGameScreen();
     }
 }
