+++ src/main/java/UI/E_LeaderboardScreen.java
@@ -8,29 +8,52 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import javax.swing.ImageIcon;
 
+/**
+ * This is the leaderboard screen, which will have two buttons
+ *
+ * one button which takes the player back to main menu
+ *
+ * one button which displays the leaderboard
+ *
+ * the leaderboard is determined by which player has the most amount of money
+ * only the top 3 players are displayed
+ */
 public class E_LeaderboardScreen extends A0_MainScreen {
-    //title name
+
+    /**
+     * initializes leaderBoard title
+     */
     JPanel leaderboardTitlePanel;
     JLabel leaderboardTitleLabel;
 
-    //backButton
+    /**
+     * initializes back button which takes player back to menu
+     */
     JPanel backButtonPanel;
     JButton backButton;
 
-    //leaderboardScreen
+    /**
+     * initializes the window that shows the leaderboard
+     */
     JPanel leaderboardTextPanel;
-    Font titleFont = new Font("Times New Roman", Font.PLAIN, 96);
 
-
+    /**
+     * ActionListener which transitions the leaderboard screen to menu screen
+     */
     leaderboardtoMainMenuHandler lmHandler = new leaderboardtoMainMenuHandler();
 
+    /**
+     * Setting the parameters for the title, back button and leaderboard
+     *
+     * adding all the panels to the container con, which was initialized in A0_MainScreen
+     */
     public E_LeaderboardScreen() {
         backButton = new JButton("Back");
 
         //makes title
         leaderboardTitlePanel = new JPanel();
-        leaderboardTitlePanel.setBounds(100, 30, 600, 50);
-        leaderboardTitlePanel.setBackground(Color.gray);
+        leaderboardTitlePanel.setBounds(100, 80, 600, 200);
+        leaderboardTitlePanel.setBackground(Color.black);
         leaderboardTitleLabel = new JLabel("Leaderboard", SwingConstants.CENTER);
         leaderboardTitleLabel.setForeground(Color.white);
         leaderboardTitleLabel.setFont(titleFont);
@@ -39,7 +62,7 @@ public class E_LeaderboardScreen extends A0_MainScreen {
         backButtonPanel = new JPanel();
         backButtonPanel.setBounds(300, 400, 200, 100);
         backButtonPanel.setBackground(Color.black);
-        backButton.setBackground(Color.black);
+        backButton.setBackground(Color.white);
         backButton.setForeground(Color.black);
         backButton.setFont(normalFont);
         backButton.addActionListener(lmHandler);
@@ -51,11 +74,9 @@ public class E_LeaderboardScreen extends A0_MainScreen {
         con.add(leaderboardTitlePanel);
         con.add(backButtonPanel);
 
-        // somewhere for the display leaderboard button to go
         leaderboardTextPanel = new JPanel();
-        leaderboardTextPanel.setBounds(300, 300, 150, 30);
-        leaderboardTitleLabel.setForeground(Color.black);
-        leaderboardTitleLabel.setFont(normalFont);
+        leaderboardTextPanel.setBounds(300, 300, 200, 50);
+        leaderboardTextPanel.setBackground(Color.black);
 
         // custom icon
         java.net.URL imgURL = E_LeaderboardScreen.class.getClassLoader().getResource("medals_icon.jpg");
@@ -74,6 +95,7 @@ public class E_LeaderboardScreen extends A0_MainScreen {
                 JOptionPane.showMessageDialog(con, info, "Current Leaderboard", JOptionPane.INFORMATION_MESSAGE, icon2); // make it work
             }
         });
+
         leaderboardTextPanel.add(btn_show_leaderboard);
         con.add(leaderboardTextPanel);
 
@@ -91,12 +113,15 @@ public class E_LeaderboardScreen extends A0_MainScreen {
     }
 
     /**
-     * To create the Leaderboard Screen
+     * initializes the leaderboard screen to be called by main
      */
     public static void createAndShowLeaderBoardScreen() {
         new E_LeaderboardScreen();
     }
 
+    /**
+     * when this screen is called, the screen is displayed
+     */
     public static void main(String[] args) {
         javax.swing.SwingUtilities.invokeLater(new Runnable() {
             public void run() {
